Fix CI lint and chat suspense
