+++ apps/backend/src/prompts/__init__.py
@@ -1,12 +1,12 @@
 """Prompts package."""
 
+from src.prompts.ai_advisor import get_ai_advisor_prompt
 from src.prompts.statement import (
     INSTITUTION_HINTS,
     SYSTEM_PROMPT,
     VALIDATION_PROMPT,
     get_parsing_prompt,
 )
-from src.prompts.ai_advisor import get_ai_advisor_prompt
 
 __all__ = [
     "SYSTEM_PROMPT",
